+++ module/logging/src/main/kotlin/sibyl/module/logging/LoggingModule.kt
@@ -57,7 +57,7 @@ class LoggingModule(
         logger.info { "loaded data source" }
     }
 
-    override fun start() {
+    override suspend fun start() {
         logsFolder.mkdirs()
     }