+++ module/reminders/src/test/resources/migrations/V1.1__reminders_test_data.sql
@@ -1,9 +1,12 @@
 INSERT INTO "sibyl-reminders".reminders (
     id,
-    timestamp,
     message,
     username,
-    userid
+    userid,
+    gateway,
+    requestedAt,
+    fulfilledat,
+    target
 ) VALUES
-( 1, '2020-08-20 06:35:24.000000', 'test message', 'testuser', 'testuser@test' )
+( 1, 'test message', 'testuser', 'testuser@test', 'test-gateway', '2020-08-20 06:35:24.000000', '2020-08-20 07:00:00.000000', '2020-08-20 07:00:05.000000')
 ;