+++ sample/src/main/kotlin/sibyl/test/ExitCommand.kt
@@ -1,8 +1,5 @@
 package sibyl.test
 
-import com.github.ajalt.clikt.parameters.arguments.argument
-import com.github.ajalt.clikt.parameters.arguments.multiple
-import com.github.ajalt.clikt.parameters.arguments.transformAll
 import sibyl.commands.SibylCommand
 import kotlin.system.exitProcess
 
@@ -11,7 +8,7 @@ class ExitCommand : SibylCommand(
     invokeWithoutSubcommand = true
 ) {
     override fun run() {
-        require(message.username == "Nikky") {"needs to be sent by admin"}
+        require(causeMessage.username == "Nikky") {"needs to be sent by admin"}
         exitProcess(0)
     }
 }