+++ RELEASE_NOTES.md
@@ -2,6 +2,8 @@
 
 ### 0.0.3
 
+- add reminders module
+- add sqldelight based postgres bindings
 - add json schema based configuration
 - add pastebin, paste.ee and hastebin modules